requirements.txt
asgiref==3.5.2
black==22.6.0
boto3==1.24.56
botocore==1.27.56
click==8.1.3
colorama==0.4.5
Django==4.1
django-storages==1.13.1
gunicorn==20.1.0
jmespath==1.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
python-dateutil==2.8.2
pytz==2022.2.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.1
urllib3==1.26.11